Add settings to adjust label origin
Add another collapsible menu to adjust the top left origin of the label for printing. This could be handy in case some label sizes are too small and we need to move it closer to the left. The label preview has also been updated to take this into account.

Also use ui.timer instead of threading.Timer for  debouncing to allow NiceGUI to render UI elements such as ui.notify, otherwise the framework complains about known slots due to trying to render UI elements in background threads.

The debouncing delay has also been increased to 1s.